简单功能拆分完成，创建单独的service和schema文件
